Split the SaisieController into multiple controllers
